+++ src/datastore/async_methods/save.js
@@ -70,7 +70,7 @@ function save(resourceName, id, options) {
       if (options.cacheResponse) {
         return _this.inject(definition.name, attrs, options);
       } else {
-        return attrs;
+        return _this.createInstance(resourceName, attrs, options);
       }
     });
 }